python_server/requirements.txt
git+https://github.com/fairtracks/fairtracks_validator@ef1791f0180e1b75bd65f779d51528a665169d9a#egg=fairtracks-validator&subdirectory=python
pyyaml
flask
flask-cors
flask-restplus
flask-compress
flup
